install_suse(): check if gpg key file dir exists
The defined location (/etc/pki/rpm-gpg) for  the gpg key file does not exists by
default in OpenSUSE.

This commit adds a check whether that directory exists, and creates it
if needed.

Making sure the specified directory exists prevents the situation where
the gpg key file cannot be written because the directory did not
exists---which subsequently results in zypper failing to read the
nonexistent key file and finally not accepting the added repository.