moving from apt-get to apt
reason for suggesting change:
 
1. newer suggested tooling
2. doesn't store cache of packages so saves space
ref: https://askubuntu.com/a/897605


I understand that the recommendation as per apt man pages is to keep apt-get instead of apt however in this specific script we are unnecessarily storing the cache using apt removes that cache. also those functional enough to understand that the cache should be clean would probably not be using this script anyways.